refactor: 金額と数量の計算をBigDecimalに寄せる

+++ projects/crypto-autotrading-app/src/main/kotlin/cryptoautotrading/domain/repository/ResultOutputPort.kt
@@ -1,6 +1,7 @@
 package cryptoautotrading.domain.repository
 
 import cryptoautotrading.domain.model.TradeAction
+import java.math.BigDecimal
 
 /**
  * 判定結果を出力するインターフェース
@@ -10,10 +11,10 @@ interface ResultOutputPort {
      * 結果を出力する
      */
     fun printResult(
-        price: Double,
+        price: BigDecimal,
         action: TradeAction,
         reason: String,
-        profitAndLoss: Double,
-        estimatedProfitAndLoss: Double
+        profitAndLoss: BigDecimal,
+        estimatedProfitAndLoss: BigDecimal
     )
 }
